% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symbols.R
\name{add_symbols}
\alias{add_symbols}
\title{add_symbols}
\usage{
add_symbols(bouts_values, bouts_lengths, f, bts)
}
\arguments{
\item{bouts_values}{A vector representing the cut-point classes of the corresponding bouts (e.g. 1 = SB, 2 = LPA, 3 = MPA, 4 = VPA)}

\item{bouts_lengths}{A vector representing the lengths (number of epochs) of the corresponding bouts}

\item{f}{An integer specifying the number of epochs per minute}

\item{bts}{A vector of integers defining the bout duration in minutes}
}
\value{
maps A vector of integers specifying a movement sequence map
}
\description{
'add_symbols' generates the corresponding sequence map from the detected segments (bout values and lengths obtained with the function 'detect_bouts') by applying the symbols as defined in Chinapaw et al (2019)
}
