% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggAccFile_self.R
\name{aggAccFile_self}
\alias{aggAccFile_self}
\title{aggAccFile_self}
\usage{
aggAccFile_self(object, by, which = "counts", x = NULL, keep.error = FALSE)
}
\arguments{
\item{object}{...}

\item{by}{...}

\item{which}{...}

\item{x}{...}

\item{keep.error}{...}
}
\value{
out
}
\description{
aggAccFile_self
}
