% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcoding_main_last.R
\name{barcoding_main_last}
\alias{barcoding_main_last}
\title{barcoding_main_last}
\usage{
barcoding_main_last(
  path_input,
  tz = "Europe/London",
  fileid = "test",
  epochsize = 15,
  which = "y",
  rescale.epoch = 15,
  minwear = 480,
  zerocounts = 60,
  cutpoints = c(0, 100, 2296, 4012),
  bts = c(0, 5, 10, 30),
  collapse.by = "\%Y-\%m-\%d",
  keep.error = FALSE,
  tolerance_function = "V1"
)
}
\arguments{
\item{path_input}{...}

\item{tz}{...}

\item{fileid}{...}

\item{epochsize}{...}

\item{which}{...}

\item{rescale.epoch}{...}

\item{minwear}{...}

\item{zerocounts}{...}

\item{cutpoints}{...}

\item{bts}{...}

\item{collapse.by}{...}

\item{keep.error}{...}

\item{tolerance_function}{...}
}
\value{
results
}
\description{
barcoding_main_last
}
