% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_nonwear.R
\name{detect_nonwear}
\alias{detect_nonwear}
\title{detect_nonwear}
\usage{
detect_nonwear(counts, zerocount, epochlengthPerMin)
}
\arguments{
\item{counts}{A vector containing the (aggregated/resampled) accelerometer counts}

\item{zerocount}{An integer that defines the non-wear time as number of consecutive epochs containing zero counts}

\item{epochlengthPerMin}{An integer that defines the number of epochs per minute}
}
\value{
counts A list that consists of \item{values}{A vector representing the cut-point classes of the corresponding bouts (e.g. 0 = non-wear, 1 = SB, 2 = LPA, 3 = MPA, 4 = VPA)} \item{lengths}{A vector representing the lengths (number of epochs) of the corresponding bouts}
}
\description{
'detect_nonwear' detects non-wear bouts and labels them as -999, without tolerance
}
