% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcode.R
\name{generate_barcode}
\alias{generate_barcode}
\title{generate_barcode}
\usage{
generate_barcode(bouts_values, bouts_lengths, f, bts)
}
\arguments{
\item{bouts_values}{...}

\item{bouts_lengths}{...}

\item{f}{...}

\item{bts}{...}
}
\value{
barcodes
}
\description{
generate_barcode
}
