% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence_map.R
\name{generate_sequence_map}
\alias{generate_sequence_map}
\title{generate_sequence_map}
\usage{
generate_sequence_map(bouts_values, bouts_lengths, f, bts)
}
\arguments{
\item{bouts_values}{A vector representing the cut-point classes of the corresponding bout lengths (e.g. 1 = SB, 2 = LPA, 3 = MPA, 4 = VPA)}

\item{bouts_lengths}{A vector representing the lengths (number of epochs) of the corresponding bout values}

\item{f}{An integer specifying the number of epochs per minute}

\item{bts}{A vector of integers that defines the bout durations in minutes}
}
\value{
maps A vector of integers specifying a movement sequence map
}
\description{
'generate_sequence_map' generates the corresponding sequence map from the bout values and lengths obtained with the function 'tolerated_bouts' using the symbols as defined in ChinaPaw et al (2019)
}
