% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_windows.R
\name{get_windows}
\alias{get_windows}
\title{get_windows}
\usage{
get_windows(windows, test)
}
\arguments{
\item{windows}{A character vector of length 2 specifying the start and end times in "HH:MM:SS" format.}

\item{test}{A data.frame object that contains the elements 'ts_agg' and 'outcome'.}
}
\value{
A data.frame object where the time stamp component of 'ts_agg' is between the specified start and end times and only the filtered elements 'outcome' and 'ts_agg' are returned.
}
\description{
'get_windows' filters the read in data based on whether the time stamp (ts_agg) is between specified start and end times
}
