% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.barcode.R
\name{long.barcode}
\alias{long.barcode}
\title{long.barcode}
\usage{
long.barcode(short_barcoding, sub_barcode)
}
\arguments{
\item{short_barcoding}{...}

\item{sub_barcode}{...}
}
\value{
short_barcoding
}
\description{
long.barcode
}
