% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{readdata}
\usage{
readdata(file, path_input, fileid, tz, sparse = FALSE, fault = 32767)
}
\arguments{
\item{file}{...}

\item{path_input}{...}

\item{fileid}{...}

\item{tz}{...}

\item{sparse}{...}

\item{fault}{...}
}
\value{
out
}
\description{
readdata
}
