% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_time_indicator.R
\name{reformat_time_indicator}
\alias{reformat_time_indicator}
\title{reformat_time_indicator}
\usage{
reformat_time_indicator(object, minn)
}
\arguments{
\item{object}{An object of class accfile}

\item{minn}{An integer, either specifying the epoch number or the timestamp in as.POSIX format(e.g., ’\%Y-\%m-\%d \%H:\%M:\%S’)}
}
\value{
ss An integer that defines the epoch number corresponding to a timestamp
}
\description{
'reformat_time_indicator' calculates the timestamp from epoch number or epoch number from timestamp
}
