% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{run_pipeline}
\usage{
run_pipeline(
  path_input,
  tz = "Europe/London",
  fileid = "test",
  epochsize = 15,
  which = "y",
  rescale.epoch = 15,
  minwear = 480,
  zerocounts = 60,
  cutpoints = c(0, 100, 2296, 4012),
  bts = c(0, 5, 10, 30),
  collapse.by = "\%Y-\%m-\%d",
  keep.error = FALSE,
  bout_algorithm = "V2"
)
}
\arguments{
\item{path_input}{A string specifying the path to the folder containing the accelerometry (ActiGraph) .csv files}

\item{tz}{A string specifying the time zone to be used for the conversion (see strptime) (Default : "Europe/London")}

\item{fileid}{A string specifying the label for the file identifier (Default : "test")}

\item{epochsize}{An integer that defines the epoch length in seconds (Default : 15)}

\item{which}{A string one of c("counts", "steps") for gt1m files or one of c("x", "y", "z", "steps") for gt3x files (Default : "y")}

\item{rescale.epoch}{An integer that defines the epoch length in seconds to which the accelerometry data needs to be resampled (e.g. 60 to aggregate the data per minute) (Default : 15). Note: it cannot be less than epochsize.}

\item{minwear}{An integer that defines the minimum wear time in minutes that constitutes a valid day (Default : 480)}

\item{zerocounts}{An integer that defines the non-wear time as number of consecutive epochs containing zero counts (Default : 60)}

\item{cutpoints}{A vector of integers that defines the cut-point threshold values in counts per minute (cpm). For example if value = c(0, 100, 2296, 4012) the corresponding thresholds are: SB 0 - 100, LPA 100 - 2296, MPA 2296 - 4012, VPA >= 4012 (Default : c(0, 100, 2296, 4012))}

\item{bts}{A vector of integers that defines the bout duration in minutes (Default : c(0, 5, 10, 30))}

\item{collapse.by}{A string specifying the format of the date in the accelerometer data (Default : "\%Y-\%m-\%d")}

\item{keep.error}{A boolean that flags whether errors should be omitted (Default : FALSE)}

\item{bout_algorithm}{One of c("V1", "V2") defining the tolerance function used for the bout calculation, where "V1" looks whether a pair of segments (based on intensity) is within a tolerance of 10 \% using the function 'xinhui_bout_algorithm'; "V2" looks at the whole of segments to identify breaks in bouts within this tolerance (Default : "V2")}
}
\value{
results A list of \item{sequence_day_level}{Contains a matrix of day level sequence maps of the data measured for each subject file} \item{sequence_recording_level}{Contains one sequence map for all data measured for each subject file}
}
\description{
'run_pipeline' returns sequence maps structured on day and recording level
}
\details{
This high-level function ties all functions together: 1) reads in the accelerometer data from .csv files using function 'read_file', 2) aggregates/resamples the accelerometer data using function 'resample', and 3) generates the sequence maps structured per day and per recording using the function 'generate_sequence'.
}
