% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_mapping_main_last.R
\name{sequence_mapping_main_last}
\alias{sequence_mapping_main_last}
\title{sequence_mapping_main_last}
\usage{
sequence_mapping_main_last(
  path_input,
  tz = "Europe/London",
  fileid = "test",
  epochsize = 15,
  which = "y",
  rescale.epoch = 15,
  minwear = 480,
  zerocounts = 60,
  cutpoints = c(0, 100, 2296, 4012),
  bts = c(0, 5, 10, 30),
  collapse.by = "\%Y-\%m-\%d",
  keep.error = FALSE,
  tolerance_function = "V2"
)
}
\arguments{
\item{path_input}{A string specifying the path to the folder containing the accelerometry (ActiGraph) .csv files}

\item{tz}{A string specifying the time zone to be used for the conversion (see strptime) (Default : "Europe/London")}

\item{fileid}{A string specifying the label for the file identifier (Default : "test")}

\item{epochsize}{An integer that defines the epoch length in seconds (Default : 15)}

\item{which}{A string one of c("counts", "steps") for gt1m files or one of c("x", "y", "z", "steps") for gt3x files (Default : "y")}

\item{rescale.epoch}{An integer that defines the epoch length in seconds to which the accelerometry data needs to be rescaled (e.g. 60 to aggregate the data per minute) (Default : 15)}

\item{minwear}{An integer that defines the minimum wear time in minutes that constitutes a valid day (Default : 480)}

\item{zerocounts}{An integer that defines the non-wear time as number of consecutive epochs containing zero counts (Default : 60)}

\item{cutpoints}{A vector of integers that defines the cut-point threshold values in counts per minute (cpm). For example if value = c(0, 100, 2296, 4012) the corresponding thresholds are: SB 0 - 100, LPA 100 - 2296, MPA 2296 - 4012, VPA >= 4012 (Default : c(0, 100, 2296, 4012))}

\item{bts}{A vector of integers that defines the bout durations in minutes (Default : c(0, 5, 10, 30))}

\item{collapse.by}{A string specifying the format of the date in the accelerometer data (Default : "\%Y-\%m-\%d")}

\item{keep.error}{A boolean that flags wheter errors should be omitted (Default : FALSE)}

\item{tolerance_function}{One of c("V1", "V2") defining the tolerance function used for the bout calculation, where "V1" looks whether a pair of segments (based on intensity) is within a tolerance of 10\% using the function 'tolerance'; "V2" looks at the whole of segments to identify breaks in bouts within this tolerance (Default : "V2")}
}
\value{
results A list of \item{short_sequence}{Contains a sequence map for each day of data measured for each subject file} \item{long_sequence}{Contains one sequence map for all data measured for each subject file}
}
\description{
'sequence_mapping_main_last' returns short and long sequence maps
}
\details{
This function reads in the accelerometer files using function 'readdata', aggregates the accelerometer data using function 'aggAccFile_self', and generates the short and long sequence maps using the function 'bouts_length_filter'.
}
