% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorting.barcode.R
\name{shorting.barcode}
\alias{shorting.barcode}
\title{shorting.barcode}
\usage{
shorting.barcode(short_barcoding, barcode_per_day)
}
\arguments{
\item{short_barcoding}{...}

\item{barcode_per_day}{...}
}
\value{
short_barcoding
}
\description{
shorting.barcode
}
