% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_per_day.R
\name{structure_per_day}
\alias{structure_per_day}
\title{structure_per_day}
\usage{
structure_per_day(day_level_mapping, map_per_day)
}
\arguments{
\item{day_level_mapping}{A vector of integers specifying a movement sequence map}

\item{map_per_day}{A vector of integers specifying a movement sequence map of one day}
}
\value{
day_level_mapping A matrix of integers with rows specifying a movement sequence map of one day
}
\description{
'structure_per_day' organizes the sequence map as a new row per day
}
