% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{tolerance}
\usage{
tolerance(
  bouts_values,
  bouts_lengths,
  allow_bout,
  timethreshold1,
  timethreshold2,
  Nepoch_per_minute
)
}
\arguments{
\item{bouts_values}{...}

\item{bouts_lengths}{...}

\item{allow_bout}{...}

\item{timethreshold1}{...}

\item{timethreshold2}{...}

\item{Nepoch_per_minute}{...}
}
\value{
tol
}
\description{
tolerance
}
