% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerated_bouts.R
\name{tolerated_bouts}
\alias{tolerated_bouts}
\title{tolerated_bouts}
\usage{
tolerated_bouts(
  bouts_values,
  bouts_lengths,
  timethresholds = c(5, 10, 30, 60),
  tolerance_class = c(4, 3, 2),
  Nepoch_per_minute,
  tolerance_function = "V1"
)
}
\arguments{
\item{bouts_values}{...}

\item{bouts_lengths}{...}

\item{timethresholds}{...}

\item{tolerance_class}{...}

\item{Nepoch_per_minute}{...}

\item{tolerance_function}{...}
}
\value{
bb
}
\description{
tolerated_bouts
}
