% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsFromEpoch_self.R
\name{tsFromEpoch_self}
\alias{tsFromEpoch_self}
\title{tsFromEpoch_self}
\usage{
tsFromEpoch_self(object, minn)
}
\arguments{
\item{object}{...}

\item{minn}{...}
}
\value{
ss
}
\description{
tsFromEpoch_self
}
