% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts_length_filter.R
\name{bouts_length_filter}
\alias{bouts_length_filter}
\title{bouts_length_filter}
\usage{
bouts_length_filter(
  counts,
  timeline,
  file_name,
  epochsize,
  minwear,
  zerocounts,
  cutpoints,
  bts,
  tz,
  tolerance_function = "V1"
)
}
\arguments{
\item{counts}{...}

\item{timeline}{...}

\item{file_name}{...}

\item{epochsize}{...}

\item{minwear}{...}

\item{zerocounts}{...}

\item{cutpoints}{...}

\item{bts}{...}

\item{tz}{...}

\item{tolerance_function}{...}
}
\value{
result
}
\description{
bouts_length_filter
}
