% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelNonWear.R
\name{labelNonWear}
\alias{labelNonWear}
\title{labelNonWear}
\usage{
labelNonWear(counts, zerocount, epochlengthPerMin)
}
\arguments{
\item{counts}{...}

\item{zerocount}{...}

\item{epochlengthPerMin}{...}
}
\value{
counts
}
\description{
labelNonWear
}
