% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_per_recording.R
\name{structure_per_recording}
\alias{structure_per_recording}
\title{structure_per_recording}
\usage{
structure_per_recording(day_level_mapping, day_level_map)
}
\arguments{
\item{day_level_mapping}{A vector of integers specifying a movement sequence map}

\item{day_level_map}{A vector of integers specifying the movement sequence map of one day}
}
\value{
day_level_mapping A vector of integers specifying the movement sequence map of all (valid) days
}
\description{
'structure_per_recording' reformats day level sequence maps a into recording level sequence map
}
