% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerated_bouts.R
\name{tolerated_bouts}
\alias{tolerated_bouts}
\title{tolerated_bouts}
\usage{
tolerated_bouts(
  bouts_values,
  bouts_lengths,
  timethresholds = c(5, 10, 30, 60),
  tolerance_class = c(4, 3, 2),
  Nepoch_per_minute,
  tolerance_function = "V1"
)
}
\arguments{
\item{bouts_values}{A vector representing the cut-point intensity classes of the corresponding bouts (e.g. 1 = SB, 2 = LPA, 3 = MPA, 4 = VPA)}

\item{bouts_lengths}{A vector representing the lengths (number of epochs) of the corresponding bouts}

\item{timethresholds}{A vector specifying the upper thresholds of the bout durations in minutes (Default : c(5, 10, 30, 60))}

\item{tolerance_class}{A vector specifying the order of the cut-point intensity classes for which tolerance is allowed (Default : c(4, 3, 2))}

\item{Nepoch_per_minute}{An integer defining the numer of epochs per minute}

\item{tolerance_function}{One of c("V1", "V2") defining the tolerance function used for the bout tolerance, where "V1" looks whether a pair of segments (based on intensity) is within a tolerance of 10\% using the function 'tolerance'; "V2" looks at the whole of segments to identify breaks in bouts within this tolerance (Default : "V2")}
}
\value{
bb A list of: \item{values}{A vector with the updated bout values} \item{lengths}{A vector with the updated bout lengths}
}
\description{
'tolerated_bouts' identifies the bouts of which the length lie within a tolerance of 10\% below the lower threshold with a maximum of three consecutive minutes and updates the vector of bouts accordingly
}
