% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsFromEpoch_self.R
\name{tsFromEpoch_self}
\alias{tsFromEpoch_self}
\title{tsFromEpoch_self}
\usage{
tsFromEpoch_self(object, minn)
}
\arguments{
\item{object}{An object of class accfile}

\item{minn}{An integer, either specifying the epoch number or the timestamp in aPOSIXformat(e.g., ’\%Y-\%m-\%d \%H:\%M:\%S’)}
}
\value{
ss An integer that defines the epoch number  corresponding  to  a timestamp
}
\description{
'tsFromEpoch_self' calculates the timestamp from epoch number or epoch number from timestamp
}
